+++ NEWS.md
@@ -1,6 +1,8 @@
 ## 2.0.2 ##
 * Update exclusion documentation to include line_exclusions and function
   exclusions (#191).
+* Explicitly set default packages (including methods) (#183, #180)
+* Set R_LIBS and R_LIBS_SITE as well as R_LIBS_USER (#188).
 * Automatically exclude RcppExport files (#170).
 * Memoised and Vectorized functions now able to be tracked.